+++ man-roxygen/return-ml-estimation.R
@@ -5,7 +5,7 @@
 #' * `coefficients` : A named vector of estimated coefficients (parameters of the
 #'   assumed distribution). **Note**: The parameters are given in the
 #'   (log-)location-scale-parameterization.
-#' * `confint` : Confidence intervals for (log-)location-scale parameters.
+#' * `confint` : Confidence intervals for the (log-)location-scale parameters.
 #' * `shape_scale_coefficients` : Only included if `distribution` is `"weibull"`
 #'   or `"weibull3"` (parameterization used in [Weibull][stats::Weibull]).
 #' * `shape_scale_confint` : Only included if `distribution` is `"weibull"`