+++ NEWS.md
@@ -18,6 +18,7 @@
 ## New Features
 
 ### Parametric Models
+* `loglik_profiling()` is now an S3 generic. `loglik_profiling()` becomes `loglik_profiling.default()`. Added `loglik_profiling.wt_reliability_data()`.
 * `loglik_function()` is now an S3 generic. `loglik_function()` becomes `loglik_function.default()`. Added `loglik_function.wt_reliability_data()`.
 * `rank_regression()`: New arguments `direction` (specifies direction of dependency in the model) and `control` (enables access to argument `control` in `optim()`).  
 * `r_squared_profiling()`: New argument `direction`.